feat:예약 페이지 하드코딩제거
